+++ src/main/java/org/services/IService.java
@@ -9,7 +9,7 @@ public interface IService<D extends Model>{
 
     D create(Map<String, Object> data) throws Exception;
 
-    D getObjectById(int id) throws Exception;
+    D getObjectById(long id) throws Exception;
 
-    D delete(int id) throws Exception;
+    D delete(long id) throws Exception;
 }
